Brainstorming/pyproject.toml
[project]
name = "brainstorming"
version = "0.0.1"
description = "Squad-AI Brainstorming tools — TTS, hooks, and experiments."
authors = [{ name = "Etienne Millerioux" }]
readme = "README.md"
keywords = ['python', 'tts', 'claude-code']
requires-python = ">=3.10,<4.0"
dependencies = [
    "google-cloud-texttospeech>=2.21.0",
    "sounddevice>=0.5.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
]
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.urls]
Homepage = "https://fpgmaas.github.io/brainstorming/"
Repository = "https://github.com/fpgmaas/brainstorming"
Documentation = "https://fpgmaas.github.io/brainstorming/"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pre-commit>=4.5.1",
    "tox-uv>=1.29.0",
    "deptry>=0.24.0",
    "mypy>=1.19.1",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.14",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
    "mkdocstrings[python]>=1.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["brainstorming"]

[tool.mypy]
files = ["brainstorming"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["brainstorming"]

